apps/realty-main/requirements.compiled.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.compiled.txt requirements.in
#
asgiref==3.9.1
    # via
    #   django
    #   django-cors-headers
brotli==1.1.0
    # via whitenoise
dj-database-url==3.0.1
    # via environs
dj-email-url==1.0.6
    # via environs
django==5.2.5
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cache-url==3.4.5
    # via environs
django-cors-headers==4.7.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
environs[django]==14.3.0
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
marshmallow==4.0.0
    # via environs
packaging==25.0
    # via gunicorn
prometheus-client==0.22.1
    # via -r requirements.in
psycopg[binary]==3.2.9
    # via -r requirements.in
psycopg-binary==3.2.9
    # via psycopg
pyjwt==2.10.1
    # via djangorestframework-simplejwt
python-dotenv==1.1.1
    # via environs
sqlparse==0.5.3
    # via django
tzdata==2025.2
    # via
    #   django
    #   psycopg
whitenoise[brotli]==6.9.0
    # via -r requirements.in
